breaking out the buttons from html into js and its value into its own array in its own file
