+++ index.html
@@ -12,37 +12,16 @@
     <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no" />
     <link href="https://fonts.googleapis.com/css?family=Montserrat" rel="stylesheet" />
     <link rel="stylesheet" href="style.css" />
-    <script src="script.js" defer></script>
-    <script src="sw.js" defer></script>
-    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no" />
     <link href="https://fonts.googleapis.com/css?family=Montserrat" rel="stylesheet" />
+    <script type="module" src="./scripts/script.mjs"></script>
+    <script src="sw.js" defer></script>
 
 </head>
 
 <body>
     <h1>Ns' Taste of Bacolod Menu</h1>
     <div class="container">
-        <div class="button-container">
-            <button class="button" value="118">Pechopak (without Rice)</button>
-            <button class="button" value="128">Pechopak (with Rice)</button>
-            <button class="button" value="98">Paa (Without rice)</button>
-            <button class="button" value="108">Paa (With rice)</button>
-            <button class="button" value="38">Chicken Wings / PC (Without Rice) </button>
-            <button class="button" value="88">Chicken Wing 2 PC (With Rice)</button>
-            <button class="button" value="32">Pork Barbeque / PC (Without rice)</button>
-            <button class="button" value="78">Pork Barbeque 2 PC (With Rice) </button>
-            <button class="button" value="118">Grilled bangus (Without Rice) </button>
-            <button class="button" value="128">Grilled Bangus (with Rice)</button>
-            <button class="button" value="28">Pork Masskara / PC (Without rice)</button>
-            <button class="button" value="68">Pork Masskara 2 PC (With rice)</button>
-            <button class="button" value="18">Isaw / PC (Without Rice) </button>
-            <button class="button" value="48">Isaw 2 PC (With Rice)</button>
-            <button class="button" value="32">Atay / PC (Without rice)</button>
-            <button class="button" value="78">Atay 2 PC (With Rice) </button>
-            <button class="button" value="299">Inasal Pechopak Overload </button>
-            <button class="button" value="299">Inasal Paa Overload</button>
-            <button class="button" value="15">Extra Rice - Plain </button>
-            <button class="button" value="18">Extra Rice - Java </button>
+        <div class="button-container" id="button-container">
         </div>
