+++ scripts/arrays.mjs
@@ -1,0 +1,22 @@
+export const buttonArray = [
+    { "name": "Pechopak (without Rice)", "price": 120 },
+    { "name": "Pechopak (with Rice)", "price": 128 },
+    { "name": "Paa (Without rice)", "price": 98 },
+    { "name": "Paa (With rice)", "price": 108 },
+    { "name": "Chicken Wings / PC (Without Rice)", "price": 38 },
+    { "name": "Chicken Wing 2 PC (With Rice)", "price": 88 },
+    { "name": "Pork Barbeque / PC (Without rice)", "price": 32 },
+    { "name": "Pork Barbeque 2 PC (With Rice)", "price": 78 },
+    { "name": "Grilled bangus (Without Rice)", "price": 118 },
+    { "name": "Grilled Bangus (with Rice)", "price": 128 },
+    { "name": "Pork Masskara / PC (Without rice)", "price": 28 },
+    { "name": "Pork Masskara 2 PC (With rice)", "price": 68 },
+    { "name": "Isaw / PC (Without Rice)", "price": 18 },
+    { "name": "Isaw 2 PC (With Rice)", "price": 48 },
+    { "name": "Atay / PC (Without rice)", "price": 32 },
+    { "name": "Atay 2 PC (With Rice)", "price": 78 },
+    { "name": "Inasal Pechopak Overload", "price": 299 },
+    { "name": "Inasal Paa Overload", "price": 299 },
+    { "name": "Extra Rice - Plain", "price": 15 },
+    { "name": "Extra Rice - Java", "price": 18 }
+]
